+++ hw-ledger-transports/node.ts
@@ -4,7 +4,7 @@
 import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.27.1';
 import type { TransportDef } from './types.ts';
 
-import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.27.1';
+import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.27.2';
 
 export { packageInfo } from './packageInfo.ts';